refactor: 💡 OAuthType을 auth 패키지 밑으로  이동
✅ Closes: #68
