+++ src/main/java/com/postsquad/scoup/web/auth/controller/response/SocialAuthenticationResponseMapper.java
@@ -1,6 +1,6 @@
 package com.postsquad.scoup.web.auth.controller.response;
 
-import com.postsquad.scoup.web.auth.property.OAuthType;
+import com.postsquad.scoup.web.auth.OAuthType;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.Mappings;
